requirements.txt
faiss-gpu
numpy
pandas
scikit-learn
requests
feature-engine
optuna
xgboost
joblib
tqdm